feat: optimize UDP performance with route caching and checksum fixes
- **UDP Optimization**: Implemented route caching in `UDPEndpoint` to avoid expensive per-packet routing table lookups.
- **Checksum Fix**: Fixed a potential integer overflow in `internetChecksum` by casting operands to u32 before shifting.
- **Performance**: Added stack-allocated view optimization for single-iovec writes to reduce allocator overhead.
- **Iperf Example**: Enhanced UDP load generation with better batching, MTU-aware packet sizing, and retry logic for `WouldBlock` events.

+++ src/header.zig
@@ -196,7 +196,7 @@ pub fn internetChecksum(data: []const u8, initial: u32) u32 {
     var sum: u32 = initial;
     var i: usize = 0;
     while (i + 1 < data.len) : (i += 2) {
-        sum += (@as(u16, data[i]) << 8) | data[i + 1];
+        sum += (@as(u32, data[i]) << 8) | data[i + 1];
     }
     if (i < data.len) {
         sum += @as(u32, data[i]) << 8;
